ocr-server/pyproject.toml
[tool.poetry]
name = "ocr-server"
version = "0.1.0"
description = ""
authors = ["tjdud2 <tjdud2249@naver.com>"]
readme = "README.md"
packages = [{include = "ocr_server"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.3"
djangorestframework = "^3.14.0"
mysqlclient = "^2.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
